Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scripthash-basic-blocked.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Security-Policy" content="script-src 'sha256-0WwzeJrO6lcDUe7o6BR3lx0b8uiBvXBX5MNFFKF7iYE'">
        <script>
            if (window.testRunner)
                testRunner.dumpAsText();
            alert('PASS (1/1)');
        </script>
        <script>
        alert('FAIL (1/4)');
        </script>
        <script>alert('FAIL (2/4)');
        </script>
        <script>alert('FAIL (3/4)');</script>
<script>alert('FAIL (4/4)');</script>
    </head>
    <body>
        <p>
            This tests the effect of a valid script-hash value, with one valid
            script and several invalid ones. It passes if one alert is executed
            and four console warings are visible.
        </p>
    </body>
</html>
